+++ src/sections/Download.vue
@@ -2,6 +2,7 @@
 import { computed, type ComputedRef, ref } from 'vue'
 
 import Output from '../components/Output.vue'
+import SubSectionBorder from '../components/SubSectionBorder.vue'
 
 import type { Format, Organisation } from '../types/app'
 import type {
@@ -97,7 +98,7 @@ let distributionOption: ComputedRef<DistributionOption> = computed(() => {
 </script>
 
 <template>
-  <div class="p-2 space-y-2 border-2 border-gray-400">
+  <SubSectionBorder class="space-y-2">
     <form>
       <input
         class="file:py-1 file:px-2 file:text-xs file:border file:bg-white file:border-black file:hover:bg-gray-100 file:shadow file:cursor-pointer"
@@ -107,5 +108,5 @@ let distributionOption: ComputedRef<DistributionOption> = computed(() => {
       />
     </form>
     <Output :data="distributionOption"></Output>
-  </div>
+  </SubSectionBorder>
 </template>