+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Refactoring spacing classes to a consistent set of values
 - Refactoring two column layout into a `TwoColumn` component
 - Refactoring colour classes to a consistent set of values
+- Refactoring sub-section wrappers into a `SubSectionBorder` component
 
 ## [0.4.0] - 2023-12-27